Modified the git-secret policy -
The policy is creating a git-secret on the remote cluster of type:
Opaque. For pipelines to use the secret we need to create it with the
type: kubernetes.io/basic-auth.

When it gets generated on the hub cluster, it (the secret) is generated
correctly, but incorrectly on the target (spoke) cluster.